Bugfix: Make record id string parsing never fail early (#3144)
